Sri-Lankan-Tourism-Culture-Assistance-main/Tourism&Culture_Assistance/requirements.txt
streamlit
langchain
weaviate-client
pypdf
pandas
python-dotenv
sentence-transformers
langchain_mistralai
langchain_community